+++ frontend/src/app/page.tsx
@@ -3,13 +3,10 @@
 import { useEffect, useState, useRef, useCallback } from 'react';
 import { useRouter } from 'next/navigation';
 import { useUser } from '@/context/UserContext';
-import {
-  Network, Sparkles, FilePlus2, Brain, CalendarClock, Users,
-  PenSquare, X, Search, ChevronDown, XCircle,
-} from 'lucide-react';
+import { Network, Sparkles, FilePlus2, Brain, CalendarClock, Users, PenSquare } from 'lucide-react';
+import OnboardingModal, { type OnboardingModalHandle } from '@/components/OnboardingModal';
 
 const SCRAMBLE_CHARS = "ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789!<>-_\\/[]{}=+*^?#_";
-const API_URL = process.env.NEXT_PUBLIC_API_URL ?? 'http://localhost:5000';
 
 export default function LandingPage() {
   const router = useRouter();
@@ -18,12 +15,6 @@ export default function LandingPage() {
   const [heroMounted, setHeroMounted] = useState(false);
   const [heroText1, setHeroText1] = useState('');
   const [heroText2, setHeroText2] = useState('');
-  const [modalOpen, setModalOpen] = useState(false);
-  const [modalActive, setModalActive] = useState(false);
-  const [modalMode, setModalMode] = useState<'signin' | 'signup'>('signup');
-  const [modalStep, setModalStep] = useState(1);
-  const [classChips, setClassChips] = useState<string[]>([]);
-  const [classInput, setClassInput] = useState('');
 
   const canvasRef = useRef<HTMLCanvasElement>(null);
   const heroContentRef = useRef<HTMLDivElement>(null);
@@ -37,22 +28,13 @@ export default function LandingPage() {
   const ambientGlowRef = useRef<HTMLDivElement>(null);
   const parallaxYRef = useRef(0);
   const mouseRef = useRef({ x: 0, y: 0 });
-  const modalCloseTimeout = useRef<NodeJS.Timeout | null>(null);
+  const modalRef = useRef<OnboardingModalHandle>(null);
 
   useEffect(() => {
     if (userReady && isAuthenticated) {
       router.replace('/dashboard');
     }
   }, [userReady, isAuthenticated, router]);
-
-  useEffect(() => {
-    return () => {
-      if (modalCloseTimeout.current) {
-        clearTimeout(modalCloseTimeout.current);
-        modalCloseTimeout.current = null;
-      }
-    };
-  }, []);
 
   const scrambleText = useCallback((setter: (v: string) => void, final: string, duration: number) => {
     const start = Date.now();
@@ -216,6 +198,31 @@ export default function LandingPage() {
       updateNavChrome(sy);
       updateAmbientGlow(sy);
       lastSy = sy;
+
+      // Sticky scroll cinema
+      const sec = stickyRef.current;
+      if (sec) {
+        const rect = sec.getBoundingClientRect();
+        const progress = Math.max(0, Math.min(1, -rect.top / (rect.height - window.innerHeight)));
+        const s1 = step1Ref.current, s2 = step2Ref.current, s3 = step3Ref.current, sn = stepNumRef.current;
+        if (s1 && s2 && s3 && sn) {
+          const activate = (el: HTMLDivElement, num: string) => {
+            el.classList.remove('opacity-0', 'pointer-events-none', 'translate-y-[60px]', '-translate-y-[60px]');
+            sn.textContent = num;
+          };
+          const deactivateTop = (el: HTMLDivElement) => {
+            el.classList.add('opacity-0', 'pointer-events-none', '-translate-y-[60px]');
+            el.classList.remove('translate-y-[60px]');
+          };
+          const deactivateBottom = (el: HTMLDivElement) => {
+            el.classList.add('opacity-0', 'pointer-events-none', 'translate-y-[60px]');
+            el.classList.remove('-translate-y-[60px]');
+          };
+          if (progress < 0.33) { activate(s1, '01'); deactivateBottom(s2); deactivateBottom(s3); }
+          else if (progress < 0.66) { deactivateTop(s1); activate(s2, '02'); deactivateBottom(s3); }
+          else { deactivateTop(s1); deactivateTop(s2); activate(s3, '03'); }
+        }
+      }
     };
 
     document.addEventListener('mousemove', onMouse);
@@ -276,37 +283,6 @@ export default function LandingPage() {
     return () => cards.forEach(c => c.removeEventListener('mousemove', handler));
   }, []);
 
-  // Sticky scroll cinema
-  useEffect(() => {
-    function onScroll() {
-      const sec = stickyRef.current;
-      if (!sec) return;
-      const rect = sec.getBoundingClientRect();
-      const progress = Math.max(0, Math.min(1, -rect.top / (rect.height - window.innerHeight)));
-      const s1 = step1Ref.current, s2 = step2Ref.current, s3 = step3Ref.current, sn = stepNumRef.current;
-      if (!s1 || !s2 || !s3 || !sn) return;
-
-      function activate(el: HTMLDivElement, num: string) {
-        el.classList.remove('opacity-0', 'pointer-events-none', 'translate-y-[60px]', '-translate-y-[60px]');
-        sn!.textContent = num;
-      }
-      function deactivateTop(el: HTMLDivElement) {
-        el.classList.add('opacity-0', 'pointer-events-none', '-translate-y-[60px]');
-        el.classList.remove('translate-y-[60px]');
-      }
-      function deactivateBottom(el: HTMLDivElement) {
-        el.classList.add('opacity-0', 'pointer-events-none', 'translate-y-[60px]');
-        el.classList.remove('-translate-y-[60px]');
-      }
-
-      if (progress < 0.33) { activate(s1, '01'); deactivateBottom(s2); deactivateBottom(s3); }
-      else if (progress < 0.66) { deactivateTop(s1); activate(s2, '02'); deactivateBottom(s3); }
-      else { deactivateTop(s1); deactivateTop(s2); activate(s3, '03'); }
-    }
-    window.addEventListener('scroll', onScroll);
-    return () => window.removeEventListener('scroll', onScroll);
-  }, []);
-
   function startCounters(container: HTMLElement) {
     container.querySelectorAll<HTMLElement>('.counter, .counter-float').forEach(el => {
       if (el.classList.contains('counted')) return;
@@ -324,35 +300,6 @@ export default function LandingPage() {
       requestAnimationFrame(update);
     });
   }
-
-  function openModal(mode: 'signin' | 'signup') {
-    if (modalCloseTimeout.current) {
-      clearTimeout(modalCloseTimeout.current);
-      modalCloseTimeout.current = null;
-    }
-    setModalMode(mode);
-    setModalStep(1);
-    setModalOpen(true);
-    if (typeof window !== 'undefined') {
-      window.requestAnimationFrame(() => setModalActive(true));
-    } else {
-      setModalActive(true);
-    }
-  }
-  function closeModal() {
-    setModalActive(false);
-    if (modalCloseTimeout.current) {
-      clearTimeout(modalCloseTimeout.current);
-    }
-    modalCloseTimeout.current = setTimeout(() => {
-      setModalOpen(false);
-      modalCloseTimeout.current = null;
-    }, 420);
-  }
-  function handleGoogleContinue() { window.location.href = `${API_URL}/api/auth/google`; }
-  function handleEmailContinue() { if (modalMode === 'signin') closeModal(); else setModalStep(2); }
-  function addClass() { const v = classInput.trim(); if (v) { setClassChips(prev => [...prev, v]); setClassInput(''); } }
-  function addSuggested(name: string) { setClassChips(prev => [...prev, name]); }
 
   if (userReady && isAuthenticated) return null;
 
@@ -404,8 +351,8 @@ export default function LandingPage() {
             <span style={{ fontFamily: "var(--font-spectral), 'Spectral', Georgia, serif", fontWeight: 700, fontSize: '20px', color: '#1a5c2a', letterSpacing: '-0.02em', lineHeight: 1.1 }}>Sapling</span>
           </div>
           <div className="flex items-center">
-            <button onClick={() => openModal('signin')} className="text-[var(--brand-text2)] hover:text-[var(--brand-text1)] font-medium text-sm tracking-wide transition-all duration-300 mr-6 hidden sm:block">Sign In</button>
-            <button onClick={() => openModal('signup')} className="relative overflow-hidden group bg-[#1B6C42] text-white px-7 py-2.5 rounded-full font-medium text-sm tracking-wide shadow-sm hover:shadow-md transition-all duration-400 hover:scale-[1.04] active:scale-[0.97] landing-btn-shimmer">
+            <button onClick={() => modalRef.current?.open('signin')} className="text-[var(--brand-text2)] hover:text-[var(--brand-text1)] font-medium text-sm tracking-wide transition-all duration-300 mr-6 hidden sm:block">Sign In</button>
+            <button onClick={() => modalRef.current?.open('signup')} className="relative overflow-hidden group bg-[#1B6C42] text-white px-7 py-2.5 rounded-full font-medium text-sm tracking-wide shadow-sm hover:shadow-md transition-all duration-400 hover:scale-[1.04] active:scale-[0.97] landing-btn-shimmer">
               Get Started
             </button>
           </div>
@@ -520,7 +467,7 @@ export default function LandingPage() {
             transform: heroMounted ? 'translateY(0)' : 'translateY(25px)',
             transition: 'all 700ms cubic-bezier(0.22,1,0.36,1) 700ms',
           }} className="flex flex-col sm:flex-row gap-4 mt-10 items-center justify-center">
-            <button onClick={() => openModal('signup')} className="relative overflow-hidden group bg-[#1B6C42] text-white px-10 py-4 rounded-full font-medium text-base tracking-wide shadow-md hover:shadow-lg hover:bg-[#155A35] transition-all duration-500 hover:scale-[1.03] active:scale-[0.98] landing-btn-shimmer">
+            <button onClick={() => modalRef.current?.open('signup')} className="relative overflow-hidden group bg-[#1B6C42] text-white px-10 py-4 rounded-full font-medium text-base tracking-wide shadow-md hover:shadow-lg hover:bg-[#155A35] transition-all duration-500 hover:scale-[1.03] active:scale-[0.98] landing-btn-shimmer">
               Get Started
             </button>
             <button onClick={() => document.getElementById('features')?.scrollIntoView({ behavior: 'smooth' })} className="liquid-glass-subtle text-[var(--brand-text2)] hover:text-[var(--brand-text1)] px-10 py-4 rounded-full font-medium text-base transition-all duration-500 hover:scale-[1.02] active:scale-[0.98]">
@@ -625,7 +572,7 @@ export default function LandingPage() {
           </h2>
           <p className="text-[var(--brand-text2)] text-lg mt-6 font-light">Join students who learn smarter, not harder.</p>
           <div className="mt-10 flex flex-col items-center">
-            <button onClick={() => openModal('signup')} className="relative overflow-hidden group bg-[#1B6C42] text-white px-10 py-4 rounded-full font-medium text-base tracking-wide shadow-md hover:shadow-lg hover:bg-[#155A35] transition-all duration-500 hover:scale-[1.03] active:scale-[0.98] landing-btn-shimmer">
+            <button onClick={() => modalRef.current?.open('signup')} className="relative overflow-hidden group bg-[#1B6C42] text-white px-10 py-4 rounded-full font-medium text-base tracking-wide shadow-md hover:shadow-lg hover:bg-[#155A35] transition-all duration-500 hover:scale-[1.03] active:scale-[0.98] landing-btn-shimmer">
               Get Started
             </button>
           </div>
@@ -653,124 +600,7 @@ export default function LandingPage() {
         </div>
       </footer>
 
-      {/* ═══ Onboarding Modal ═══ */}
-      {modalOpen && (
-        <div className="fixed inset-0 z-[100] flex items-center justify-center">
-          <div
-            className={`absolute inset-0 bg-[var(--brand-text1)]/25 transition-opacity duration-400 ${modalActive ? 'opacity-100' : 'opacity-0'}`}
-            onClick={closeModal}
-          />
-          <div className={`relative w-full max-w-md mx-4 sm:mx-auto transition-all duration-400 ${modalActive ? 'opacity-100 translate-y-0 scale-100' : 'opacity-0 translate-y-4 scale-[0.97]'}`}>
-            <div className="rounded-3xl p-10 relative overflow-hidden shadow-2xl" style={{ background: 'var(--bg-panel, #f8fbf8)' }}>
-              <button onClick={closeModal} className="absolute top-5 right-5 text-[#e11d48] hover:text-[#be123c] transition-colors z-20">
-                <X className="w-6 h-6" strokeWidth={2.4} />
-              </button>
-
-              {/* Progress Dots */}
-              {modalMode === 'signup' && (
-                <div className="flex justify-center gap-2 mb-8 relative z-10">
-                  <div className={`w-12 h-1.5 rounded-full transition-colors duration-400 ${modalStep >= 1 ? 'bg-[#1B6C42]' : 'bg-[#d7dfd0]'}`} />
-                  <div className={`w-12 h-1.5 rounded-full transition-colors duration-400 ${modalStep >= 2 ? 'bg-[#1B6C42]' : 'bg-[#d7dfd0]'}`} />
-                  <div className={`w-12 h-1.5 rounded-full transition-colors duration-400 ${modalStep >= 3 ? 'bg-[#1B6C42]' : 'bg-[#d7dfd0]'}`} />
-                </div>
-              )}
-
-              <div className="relative w-full h-[320px]">
-                {/* Step 1: Auth */}
-                <div className={`absolute inset-0 w-full transition-all duration-300 ease-in-out flex flex-col justify-center ${modalStep === 1 ? 'translate-x-0 opacity-100' : '-translate-x-[30px] opacity-0 pointer-events-none'}`}>
-                  <h3 className="font-playfair text-2xl font-semibold text-[var(--brand-text1)] text-center tracking-tight">
-                    {modalMode === 'signin' ? 'Welcome back' : 'Welcome to Sapling'}
-                  </h3>
-                  <p className="text-[var(--brand-text2)] text-center text-sm mt-2 font-light">
-                    {modalMode === 'signin' ? 'Sign in to continue' : 'Start your learning journey'}
-                  </p>
-                  <button onClick={handleGoogleContinue} className="mt-8 bg-white border border-[#d9e2d7] text-[var(--brand-text1)] w-full py-3.5 rounded-xl flex items-center justify-center gap-3 shadow-sm transition-all duration-300 text-sm font-medium hover:border-[#1B6C42] hover:shadow-md focus:outline-none focus:ring-2 focus:ring-[#1B6C42]/20">
-                    <svg className="w-4 h-4" viewBox="0 0 24 24"><path d="M22.56 12.25c0-.78-.07-1.53-.2-2.25H12v4.26h5.92c-.26 1.37-1.04 2.53-2.21 3.31v2.77h3.57c2.08-1.92 3.28-4.74 3.28-8.09z" fill="#4285F4" /><path d="M12 23c2.97 0 5.46-.98 7.28-2.66l-3.57-2.77c-.98.66-2.23 1.06-3.71 1.06-2.86 0-5.29-1.93-6.16-4.53H2.18v2.84C3.99 20.53 7.7 23 12 23z" fill="#34A853" /><path d="M5.84 14.09c-.22-.66-.35-1.36-.35-2.09s.13-1.43.35-2.09V7.07H2.18C1.43 8.55 1 10.22 1 12s.43 3.45 1.18 4.93l2.85-2.22.81-.62z" fill="#FBBC05" /><path d="M12 5.38c1.62 0 3.06.56 4.21 1.64l3.15-3.15C17.45 2.09 14.97 1 12 1 7.7 1 3.99 3.47 2.18 7.07l3.66 2.84c.87-2.6 3.3-4.53 6.16-4.53z" fill="#EA4335" /></svg>
-                    Continue with Google
-                  </button>
-                  <div className="flex items-center gap-4 my-6">
-                    <div className="flex-1 h-px landing-divider opacity-80" />
-                    <span className="text-[var(--brand-text2)] text-[11px] font-medium uppercase">or</span>
-                    <div className="flex-1 h-px landing-divider opacity-80" />
-                  </div>
-                  <input type="email" placeholder="your@email.com" className="rounded-xl px-4 py-3 text-[var(--brand-text1)] placeholder:text-[var(--brand-text2)] placeholder:opacity-60 w-full transition-all text-sm mb-3 bg-white border border-[#d7dfd0] shadow-inner focus:border-[#1B6C42] focus:ring-2 focus:ring-[#1B6C42]/20 focus:outline-none" />
-                  <button onClick={handleEmailContinue} className="w-full bg-[#1B6C42] text-white py-3.5 rounded-xl font-medium text-sm hover:bg-[#155A35] shadow-sm transition-all duration-300">
-                    Continue
-                  </button>
-                  <p className="text-center text-[var(--brand-text2)] text-xs mt-6">
-                    {modalMode === 'signin' ? "Don't have an account? " : 'Already have an account? '}
-                    <button onClick={() => setModalMode(modalMode === 'signin' ? 'signup' : 'signin')} className="text-[#1B6C42] font-medium hover:underline cursor-pointer">
-                      {modalMode === 'signin' ? 'Get started' : 'Sign in'}
-                    </button>
-                  </p>
-                </div>
-
-                {/* Step 2: School Info */}
-                <div className={`absolute inset-0 w-full transition-all duration-300 ease-in-out flex flex-col justify-center ${modalStep === 2 ? 'translate-x-0 opacity-100' : 'translate-x-[30px] opacity-0 pointer-events-none'}`}>
-                  <h3 className="font-playfair text-2xl font-semibold text-[var(--brand-text1)] text-center tracking-tight">About You</h3>
-                  <p className="text-[var(--brand-text2)] text-center text-sm mt-2 mb-8 font-light">Help us personalize your experience</p>
-                  <div className="relative mb-4">
-                    <Search className="absolute left-4 top-1/2 -translate-y-1/2 text-[var(--brand-text2)] opacity-70 w-4 h-4" />
-                    <input type="text" placeholder="Search your university..." className="glass-input rounded-xl pl-10 pr-4 py-3 text-[var(--brand-text1)] placeholder:text-[var(--brand-text2)] placeholder:opacity-60 w-full transition-all text-sm" />
-                  </div>
-                  <div className="relative mb-6">
-                    <select className="glass-input appearance-none rounded-xl px-4 py-3 text-[var(--brand-text1)] w-full transition-all text-sm cursor-pointer invalid:text-[var(--brand-text2)]" required defaultValue="">
-                      <option value="" disabled hidden>Select class year</option>
-                      <option value="freshman">Freshman</option>
-                      <option value="sophomore">Sophomore</option>
-                      <option value="junior">Junior</option>
-                      <option value="senior">Senior</option>
-                      <option value="graduate">Graduate</option>
-                      <option value="other">Other</option>
-                    </select>
-                    <ChevronDown className="absolute right-4 top-1/2 -translate-y-1/2 text-[var(--brand-text2)] opacity-70 w-4 h-4 pointer-events-none" />
-                  </div>
-                  <button onClick={() => setModalStep(3)} className="w-full bg-[#1B6C42] text-white py-3.5 rounded-xl font-medium text-sm hover:bg-[#155A35] shadow-sm transition-all duration-300">
-                    Continue
-                  </button>
-                </div>
-
-                {/* Step 3: Classes */}
-                <div className={`absolute inset-0 w-full transition-all duration-300 ease-in-out flex flex-col justify-center ${modalStep === 3 ? 'translate-x-0 opacity-100' : 'translate-x-[30px] opacity-0 pointer-events-none'}`}>
-                  <h3 className="font-playfair text-2xl font-semibold text-[var(--brand-text1)] text-center tracking-tight">Your Classes</h3>
-                  <p className="text-[var(--brand-text2)] text-center text-sm mt-2 mb-8 font-light">What are you studying this semester?</p>
-                  <div className="flex gap-2 mb-4">
-                    <input
-                      type="text"
-                      value={classInput}
-                      onChange={e => setClassInput(e.target.value)}
-                      onKeyDown={e => { if (e.key === 'Enter') { e.preventDefault(); addClass(); } }}
-                      placeholder="Add a class..."
-                      className="glass-input flex-1 rounded-xl px-4 py-3 text-[var(--brand-text1)] placeholder:text-[var(--brand-text2)] placeholder:opacity-60 transition-all text-sm"
-                    />
-                    <button onClick={addClass} className="bg-[#1B6C42]/10 hover:bg-[#1B6C42]/20 text-[#1B6C42] px-5 rounded-xl text-sm font-medium transition-all">Add</button>
-                  </div>
-                  <div className="flex flex-wrap gap-2 mb-4 min-h-[42px]">
-                    {classChips.map((chip, idx) => (
-                      <div key={`${chip}-${idx}`} className="liquid-glass-subtle rounded-full px-4 py-2 text-[var(--brand-text1)] text-sm flex items-center gap-2">
-                        {chip}
-                        <button onClick={() => setClassChips(prev => prev.filter((_, i) => i !== idx))} className="text-[var(--brand-text2)] hover:text-[var(--brand-text1)] ml-1 transition-colors">
-                          <XCircle className="w-4 h-4" />
-                        </button>
-                      </div>
-                    ))}
-                  </div>
-                  <div className="mb-8 flex items-center flex-wrap gap-2">
-                    <span className="text-[var(--brand-text2)] text-xs mr-2 font-medium">Popular:</span>
-                    {['CS 101', 'Calculus I', 'Physics I'].map(s => (
-                      <span key={s} onClick={() => addSuggested(s)} className="liquid-glass-subtle text-[var(--brand-text2)] hover:text-[var(--brand-text1)] rounded-full px-3 py-1.5 text-xs cursor-pointer transition-all">{s}</span>
-                    ))}
-                  </div>
-                  <button onClick={closeModal} className="w-full bg-[#1B6C42] text-white py-3.5 rounded-xl font-medium text-sm hover:bg-[#155A35] shadow-sm transition-all duration-300">
-                    Launch Sapling 🌱
-                  </button>
-                  <button onClick={closeModal} className="w-full text-center text-[var(--brand-text2)] hover:text-[var(--brand-text1)] text-xs mt-3 cursor-pointer transition-colors">Skip for now</button>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      )}
+      <OnboardingModal ref={modalRef} />
     </div>
   );
 }
